skillpath_complete/skillpath_complete/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic==0.40.0
pdfminer.six==20231228
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
